[SYCL][Doc] Move internal design docs (#5556)
Move all the internal design documents to a separate directory.  This
declutters the "sycl/doc" directory, making it easier to find the
end-user documentation.  It also moves several documents out of
"sycl/doc/extensions" which were not actually extension specifications.

We decided that the "sycl/doc/dev" directory should hold BKMs (best
known methods) for developers, so rename it accordingly.