[SYCL][DOC] Move SPIR-V and OpenCL extensions (#5578)
The SPIR-V and OpenCL extensions are not SYCL extensions, rather they
are extensions to underlying components used by DPC++.  Therefore, we
consider them part of the design, so move them to "doc/design".

Our policy is to remove SPIR-V and OpenCL extension specifications from
this repo once they are added to the Khronos registry.  The
"SPV_INTEL_arbitrary_precision_int" extension was already added there,
so remove it from this repo.